Fix create invoice button.
